#45  Move CreateKeyspace to DDL package
