+++ cql-ast/src/main/scala/troy/cql/ast/CQL3.scala
@@ -16,10 +16,7 @@
 
 package troy.cql.ast
 
-object Keyspace {
-  trait KeyspaceOption
-  case class Replication(options: Seq[(String, String)]) extends KeyspaceOption // TODO
-}
+import troy.cql.ast.ddl.KeyspaceName
 
 //
 //trait SchemaAlteringStatement
@@ -63,7 +60,6 @@ object ConsistencyLevel {
   case object EachQuorum extends ConsistencyLevel
 }
 
-case class KeyspaceName(name: String)
 case class TableName(keyspace: Option[KeyspaceName], table: String) {
   override def toString = keyspace.map(_.name + ".").getOrElse("") + table
 }
