+++ cql-ast/src/main/scala/troy/cql/ast/ddl/Keyspace.scala
@@ -1,0 +1,7 @@
+package troy.cql.ast.ddl
+
+object Keyspace {
+  trait KeyspaceOption
+  case class Replication(options: Seq[(String, String)]) extends KeyspaceOption // TODO
+}
+case class KeyspaceName(name: String)
